third_party/blink/web_tests/virtual/pna-iframes-warning/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-private-expected.txt

This is a testharness.js-based test.
FAIL private to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL private to local: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL private to local: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS private to local: success.
PASS private to private: no preflight required.
PASS private to public: no preflight required.
Harness: the test ran to completion.
